+++ source/isaaclab_physx/test/sim/__init__.py
@@ -2,9 +2,3 @@
 # All rights reserved.
 #
 # SPDX-License-Identifier: BSD-3-Clause
-
-"""Backward-compatibility alias: ``XformPrimView`` -> :class:`FrameView`."""
-
-from .frame_view import FrameView
-
-XformPrimView = FrameView